REF: Split misc.py into other submodules
