+++ openff/system/tests/test_buckingham.py
@@ -9,7 +9,8 @@ from openff.utilities.testing import skip_if_missing
 from scipy.constants import Avogadro
 from simtk import unit as simtk_unit
 
-from openff.system.components.misc import BuckinghamvdWHandler, OFFBioTop
+from openff.system.components.mdtraj import OFFBioTop
+from openff.system.components.nonbonded import BuckinghamvdWHandler
 from openff.system.components.potentials import Potential
 from openff.system.components.smirnoff import ElectrostaticsMetaHandler
 from openff.system.exceptions import GMXMdrunError
