+++ openff/system/tests/test_external.py
@@ -7,7 +7,7 @@ from openff.units import unit
 from simtk import openmm
 from simtk.unit import nanometer as nm
 
-from openff.system.components.misc import OFFBioTop
+from openff.system.components.mdtraj import OFFBioTop
 from openff.system.stubs import ForceField
 from openff.system.tests.base_test import BaseTest
 from openff.system.tests.energy_tests.openmm import (
