+++ openff/system/tests/test_interop/test_lammps.py
@@ -4,7 +4,7 @@ import pytest
 from openff.toolkit.topology import Molecule
 from simtk import unit as omm_unit
 
-from openff.system.components.misc import OFFBioTop
+from openff.system.components.mdtraj import OFFBioTop
 from openff.system.stubs import ForceField
 from openff.system.tests.energy_tests.lammps import (
     _write_lammps_input,
